Foi criado o modulo auth para organizar as features de autenticacao do usuario
