+++ front/app/utilidades.py
@@ -1,0 +1,3 @@
+# Define API base URL (replace with your actual FastAPI URL)
+API_PORT = "8000"
+API_URL = f"http://localhost:{API_PORT}"  # Replace with your FastAPI app's URL and port
