+++ front/templates/auth/forgot-password.html
@@ -28,7 +28,7 @@
     <div class="card-body login-card-body">
       <p class="login-box-msg">You forgot your password? Here you can easily retrieve a new password.</p>
 
-      <form action="{{ url_for('login') }}" method="get">
+      <form action="{{ url_for('auth.login') }}" method="get">
         <div class="input-group mb-3">
           <input type="email" class="form-control" placeholder="Email" name="email">
           <div class="input-group-append">
@@ -46,10 +46,10 @@
       </form>
 
       <p class="mt-3 mb-1">
-        <a href="{{ url_for('login') }}">Login</a>
+        <a href="{{ url_for('auth.login') }}">Login</a>
       </p>
       <p class="mb-0">
-        <a href="{{ url_for('register') }}" class="text-center">Register a new membership</a>
+        <a href="{{ url_for('auth.register') }}" class="text-center">Register a new membership</a>
       </p>
     </div>
     <!-- /.login-card-body -->
