+++ front/templates/base.html
@@ -70,7 +70,7 @@ scratch. This page gets rid of all links and provides the needed markup only.
       </li>
       <!-- Off -->
       <li class="nav-item">
-        <a class="nav-link"href="/logout">
+        <a class="nav-link"href="{{ url_for('auth.logout') }}">
           Sair
         </a>
       </li>
